backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.1
sqlalchemy==2.0.32
pydantic==2.9.2
pydantic-settings==2.5.2
alembic==1.13.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
PyMySQL==1.1.1
python-multipart==0.0.9
cryptography==43.0.1
bcrypt==3.2.2
email-validator==2.1.0.post1
aiofiles==24.1.0
openpyxl==3.1.5
